Changed mediawiz.py to accept list of params
Now iterates over list of command line params and renames them.
If a directory is encountered the same method is called recursively on
the filelist of the directory(only 1 level deep at the moment).

Previous functionality of supplying a directory as param is left intact.

Also added config.py to have configuration available across the project.